# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Davomat tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [superadmin, admin, employee]
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{id}/password:
    put:
      operationId: changeUserPassword
      tags: [users]
      summary: Change user password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List all work shifts
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create a work shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftRequest"
      responses:
        "201":
          description: Shift created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    put:
      operationId: updateShift
      tags: [shifts]
      summary: Update shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftRequest"
      responses:
        "200":
          description: Shift updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shift deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ChangePasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
        currentPassword:
          type: string
      required:
        - newPassword
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [superadmin, admin, employee]
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        shiftId:
          type: integer
          nullable: true
        shiftName:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [superadmin, admin, employee]
        departmentId:
          type: integer
          nullable: true
        shiftId:
          type: integer
          nullable: true
        position:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - username
        - password
        - firstName
        - lastName
        - role
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [superadmin, admin, employee]
        departmentId:
          type: integer
          nullable: true
        shiftId:
          type: integer
          nullable: true
        position:
          type: string
          nullable: true
        isActive:
          type: boolean
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        lateThresholdMinutes:
          type: integer
        workDays:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - startTime
        - endTime
        - lateThresholdMinutes
        - workDays
        - createdAt
    CreateShiftRequest:
      type: object
      properties:
        name:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        lateThresholdMinutes:
          type: integer
        workDays:
          type: array
          items:
            type: integer
      required:
        - name
        - startTime
        - endTime
        - lateThresholdMinutes
        - workDays
...